# mern-delivery-modified/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Food Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List restaurants
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantListResponse"
    post:
      operationId: createRestaurant
      tags: [restaurants]
      summary: Create restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get restaurant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
    patch:
      operationId: updateRestaurant
      tags: [restaurants]
      summary: Update restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
    delete:
      operationId: deleteRestaurant
      tags: [restaurants]
      summary: Delete restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /restaurants/{id}/stats:
    get:
      operationId: getRestaurantStats
      tags: [restaurants]
      summary: Get restaurant stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantStats"

  # ── Menu Categories ────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user or restaurant
      parameters:
        - name: restaurantId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── Reviews ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Restaurant:
      type: object
      required: [id, name, cuisine, deliveryTime, deliveryFee, rating, isOpen]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        cuisine:
          type: string
        image:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        address:
          type: string
        phone:
          type: ["string", "null"]
        deliveryTime:
          type: integer
          description: Estimated delivery time in minutes
        deliveryFee:
          type: number
        minOrder:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer
        isOpen:
          type: boolean
        ownerId:
          type: ["string", "null"]
        createdAt:
          type: string
    RestaurantInput:
      type: object
      required: [name, cuisine, address, deliveryTime, deliveryFee]
      properties:
        name:
          type: string
        description:
          type: string
        cuisine:
          type: string
        image:
          type: string
        coverImage:
          type: string
        address:
          type: string
        phone:
          type: string
        deliveryTime:
          type: integer
        deliveryFee:
          type: number
        minOrder:
          type: number
    RestaurantUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        cuisine:
          type: string
        image:
          type: string
        coverImage:
          type: string
        address:
          type: string
        phone:
          type: string
        deliveryTime:
          type: integer
        deliveryFee:
          type: number
        minOrder:
          type: number
        isOpen:
          type: boolean
    RestaurantListResponse:
      type: object
      required: [restaurants, total, page, limit]
      properties:
        restaurants:
          type: array
          items:
            $ref: "#/components/schemas/Restaurant"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    RestaurantStats:
      type: object
      required: [totalOrders, totalRevenue, avgRating, pendingOrders]
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        avgRating:
          type: number
        pendingOrders:
          type: integer
        todayOrders:
          type: integer
        todayRevenue:
          type: number

    # ── Menu Categories ──────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [menuItemId, name, price, quantity]
      properties:
        menuItemId:
          type: string
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        image:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, userId, restaurantId, restaurantName, items, status, subtotal, deliveryFee, total, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        restaurantId:
          type: string
        restaurantName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, out_for_delivery, delivered, cancelled]
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        deliveryAddress:
          type: string
        notes:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderInput:
      type: object
      required: [restaurantId, deliveryAddress]
      properties:
        restaurantId:
          type: string
        deliveryAddress:
          type: string
        notes:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, out_for_delivery, delivered, cancelled]
    OrderListResponse:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Reviews ───────────────────────────────────────────────────────────────
...